Add EventDetailsView
The new component shows more details on a separate page (i.e. new route) after clicking a specific event in the EventListView.

This commit contains two different approaches to pass dynamic routing information to a component for rendering and will be cleaned up in the next commit to only contain the recommended appraoch with highest decoupling.